Group Live Hackers News feed into the 'Feed' folder
