+++ HackrNewsApp/HackrNewsApp/App/AppDelegate.swift
