+++ HackrNewsApp/HackrNewsApp/Feed/ContentType.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+enum ContentType {
+    case topStories
+    case newStories
+    case bestStories
+}
