+++ HackrNewsApp/HackrNewsApp/Feed/LiveHackrNewPresentationAdapter.swift
@@ -1,0 +1,35 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import HackrNews
+import HackrNewsiOS
+
+final class LiveHackrNewPresentationAdapter: LiveHackrNewCellControllerDelegate {
+    private let model: LiveHackrNew
+    private let loader: HackrStoryLoader
+    private var task: HackrStoryLoaderTask?
+
+    var presenter: StoryPresenter?
+
+    init(model: LiveHackrNew, loader: HackrStoryLoader) {
+        self.model = model
+        self.loader = loader
+    }
+
+    func didRequestStory() {
+        presenter?.didStartLoadingStory(from: model)
+        task = loader.load { [weak self] result in
+            switch result {
+            case let .success(story):
+                self?.presenter?.didFinishLoadingStory(story: story)
+            case let .failure(error):
+                self?.presenter?.didFinishLoading(with: error)
+            }
+        }
+    }
+
+    func didCancelRequest() {
+        task?.cancel()
+    }
+}
