+++ HackrNewsApp/HackrNewsApp/Feed/LiveHackrNewsPresentationAdapter.swift
@@ -1,0 +1,28 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+import HackrNews
+import HackrNewsiOS
+
+final class LiveHackrNewsPresentationAdapter: LiveHackrNewsRefreshControllerDelegate {
+    private let liveHackrNewsloader: LiveHackrNewsLoader
+    var presenter: LiveHackrNewsPresenter?
+
+    init(liveHackrNewsloader: LiveHackrNewsLoader) {
+        self.liveHackrNewsloader = liveHackrNewsloader
+    }
+
+    func didRequestNews() {
+        presenter?.didStartLoadingNews()
+        liveHackrNewsloader.load { [weak self] result in
+            switch result {
+            case let .success(news):
+                self?.presenter?.didFinishLoadingNews(news: news)
+            case let .failure(error):
+                self?.presenter?.didFinishLoadingNews(with: error)
+            }
+        }
+    }
+}
